refactor: rename playerHealth.js to playerStas.js
